Treat references as values
This needs some tweaking, but we need to allow a reference to be treated as
a value when it's created to allow the new input map for invokable
to work.

This works? I think it does - will check in the login machine test, but
it should.

But it probably could be nicer