Simplify expressions
Massive amount of refactoring here to simplify expressions.

Still need to test it properly, but the joined up tests seem to work.

This makes it so that you can pass in things dynamically, and it
just resolves automatically as far as the consumer is concerned. So
the consumer doesn't have to wrap things in primitives or data values
anymore, and we can handle all the parsing on our side, making for a
better user experience.

Guards are still a bit wordy? I am almost certain we could write a
parser that can turn a string into a guard object.

Also need to make sure we can get from JSON to the domain via the DTOs.
I think that abstraction leaked before from JSON into the domain, hence the
wordy implementation before. I am hoping that we'll be able to keep the
nice implementation here and just do something cool in the serialisation
project to make it work.