Allow Guards to be parsed from Strings
Add logic to be able to parse a strongly typed guard object from a
String expression.

It probably doesn't cover everything and there's some requirements
like wrapping everything in parentheses that felt like a good idea
at the time for like, easier parsing. But actually might not be, but
we'll see.