Make values nullable
This might just be because I'm looking at the JSON right now, but
it feels like we can't necessarily guarantee the context and
events are set up as we expect, so can't guarantee they aren't null.

Better to have consumers handle this eventuality rather than throw
a null pointer exception.