Fix a stupid amount of things
Just so much naivety here, it's a lot more complicated than I thought
it was going to be originally. But I wouldn't be here writing
it if I knew how complicated it was going to be, so it's all worked
out fine.

Still some extras to make it nice for consumers, but this actually
works finally so that's neat.