Decouple UI from state machine
Previously we were running everything that happened in each UI node through
the parent state machine - which made things really slow due to the
processing mutex and run to completion semantics.

To bypass this, I introduced some local state in the UI nodes that
could hold what the UI knew was the current state, and then publish
those changes to the engine via events to update the source of truth.

This felt weird at the time to have two mechanisms for the same thing,
so as part of the decoupling of the UI from the state machine, the UI
doesn't do this anymore and handles its own internal state within the
UI system itself, publishing events to the state machine where appropriate.

This makes the whole system a lot less chatty, the UI nodes more
responsive than they would have been otherwise, and it lets us do
all the things we wanted to do with the state machine more easily and
with less boilerplate.

Adds validation at the same time because I forgot to commit before testing
the theory, and now it's a lot to unpick to make it look nice and logical
again.