Write node info as JSON
This is the first step of the two-step process of converting the nodes
into DTO nodes. Because the annotations are only available in
domain:ui-core, and kotlinx is only available in data:serialisation, we
hit an impasse on who does what.

The solution (I think) is for the domain project to write metadata about
the nodes that need to be constructed, and then the data:serialisation
project can read this metadata and write the DTO nodes themselves.