Replace mappers with generated code
This is the thing I actually wanted to do the most - replace the
manually created mappers with some generated code.

This works, it's a bit... not great, but it works. Can come and tidy it
up a bit later once we've got all the generation in place.

Still need to fix the OCP violations - shouldn't have to modify a class
every time we add a node, so will need to look at generating the
serialiser module next or doing a service locator or something like
that.