Refactor StegoProcessor
Too much AI, not enough development.

This refactor makes the processor easier to understand and follow, I
think. Cleans up some of the things that aren't needed in this project,
changed to names to be more correct and consistent, and extracted some
things out.

Test coverage should be fairly decent, that might need a little refactor
too, but we'll see.