feat: Autogenerate Kotlinx serialiser module
Did this one myself - funnily enough was a lot quicker. Might need to do
more of it myself from now on too.

Next up is the Koin module which is the last place that breaks OCP in
the core project - then probably something extra for the Android UI
project but not right now, this sidequest has been long enough.